requirements.txt
Django
psycopg2-binary
django-cors-headers
configparser==3.5.0
filelock==2.0.6
djangorestframework
markdown
django-filter
django-guardian
django-extensions
django-geojson
django-oauth-toolkit
Celery
redis
celery[redis]
schema